Remove ecbuild dependency add jedicmake.  Reorganize and cleanup entire build system.
The overall functionality, executable and library name, test names, files, directories
should all be preserved from the previous version.
