--- cmake/compiler_flags_Cray_Fortran.cmake
@@ -1,56 +1,0 @@
-# (C) Copyright 2009-2016 ECMWF.
-# 
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0. 
-# In applying this licence, ECMWF does not waive the privileges and immunities 
-# granted to it by virtue of its status as an intergovernmental organisation nor
-# does it submit to any jurisdiction.
-
-####################################################################
-# FLAGS COMMON TO ALL BUILD TYPES
-####################################################################
-set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -emf -rmoid -lhugetlbfs")
-
-if( HAVE_OMP )
-  set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -homp")
-else( )
-  set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -hnoomp")
-endif( )
-
-####################################################################
-# RELEASE FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_RELEASE "-O3 -hfp3 -hscalar3 -hvector3 -hPIC" )
-
-####################################################################
-# DEBUG FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_DEBUG   "-O0 -Gfast -Ktrap=fp" )
-
-####################################################################
-# BIT REPRODUCIBLE FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_BIT     "-O2 -hflex_mp=conservative -hadd_paren -hfp1" )
-
-####################################################################
-# LINK FLAGS
-####################################################################
-
-set( CMAKE_Fortran_LINK_FLAGS    "-Wl,-Map,loadmap" )
-
-####################################################################
-
-# Meaning of flags
-# ----------------
-# -hfp3     : Special optimisation for floating points
-# -Ktrap=fp : Abort on NaN
-# -R b      : Bounds checking
-# -hflex_mp=conservative -hfp1 : Obtain bit-reproducible results
-# -hflex_mp=intolerant -hfp2   : Obtain bit-reproducible results (also)
-# -hadd_paren : encourage left to right fp evaluation
-# -hscalarN , -hvectorN : optimisation for scalar and vectorisation
-# -homp/-hnoomp : Enable/Disable OpenMP
-# -rmoi : create compiler listing
