--- cmake/compiler_flags_GNU_Fortran.cmake
@@ -1,52 +1,0 @@
-# (C) Copyright 2009-2016 ECMWF.
-# 
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0. 
-# In applying this licence, ECMWF does not waive the privileges and immunities 
-# granted to it by virtue of its status as an intergovernmental organisation nor
-# does it submit to any jurisdiction.
-
-####################################################################
-# FLAGS COMMON TO ALL BUILD TYPES
-####################################################################
-
-if( HAVE_OMP )
-  set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -fopenmp")
-else( )
-  set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -fno-openmp")
-endif( )
-
-set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -D_REAL8_ -ffree-line-length-none")
-
-####################################################################
-# RELEASE FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_RELEASE "-O3 -funroll-all-loops -finline-functions ")
-
-####################################################################
-# DEBUG FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_DEBUG   "-O0 -g -fcheck=bounds -ffpe-trap=invalid,zero,overflow,underflow -fbacktrace" )
-
-####################################################################
-# BIT REPRODUCIBLE FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_BIT     "-O2 -funroll-all-loops -finline-functions" )
-
-####################################################################
-# LINK FLAGS
-####################################################################
-
-set( CMAKE_Fortran_LINK_FLAGS    "" )
-
-####################################################################
-
-# Meaning of flags
-# ----------------
-# -fstack-arrays     : Allocate automatic arrays on the stack (needs large stacksize!!!)
-# -funroll-all-loops : Unroll all loops
-# -fcheck=bounds     : Bounds checking
-
