--- cmake/compiler_flags_XL_Fortran.cmake
@@ -1,59 +1,0 @@
-# (C) Copyright 2009-2016 ECMWF.
-# 
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0. 
-# In applying this licence, ECMWF does not waive the privileges and immunities 
-# granted to it by virtue of its status as an intergovernmental organisation nor
-# does it submit to any jurisdiction.
-
-
-# WARNING!!
-# Following has not been thoroughly tested. Take these flags with grain of salt
-
-
-# Without following line, compile flags are appended to link flags
-set( CMAKE_Fortran_LINK_EXECUTABLE "<CMAKE_Fortran_COMPILER> <CMAKE_Fortran_LINK_FLAGS> <LINK_FLAGS> <OBJECTS> -o <TARGET> <LINK_LIBRARIES>")
-
-####################################################################
-# FLAGS COMMON TO ALL BUILD TYPES
-####################################################################
-
-if( ${HAVE_OMP} )
-  set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -qsmp=omp" )
-  set( CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS} -qsmp=omp" )
-else( )
-  set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -qsmp=noomp" )
-  set( CMAKE_Fortran_LINK_FLAGS "${CMAKE_Fortran_LINK_FLAGS} -qsmp=noomp" )
-endif( )
-
-set( CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS} -qfree=F90 -qsuffix=cpp=F90 -qextname -q64=largetype -qarch=pwr5 -g -qsource,list -qsaveopt -NS32648" )
-
-####################################################################
-# RELEASE FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_RELEASE "-O3 -qstrict" )
-
-####################################################################
-# DEBUG FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_DEBUG   "-O0" )
-
-####################################################################
-# BIT REPRODUCIBLE FLAGS
-####################################################################
-
-set( CMAKE_Fortran_FLAGS_BIT     "-O3 -qstrict" )
-
-####################################################################
-# LINK FLAGS
-####################################################################
-
-set( CMAKE_Fortran_LINK_FLAGS    "${CMAKE_Fortran_LINK_FLAGS} -b64 -bbigtoc -bmaxstack:0x800000000 -bloadmap:map -bmap:map")
-
-####################################################################
-
-# Meaning of flags
-# ----------------
-# todo
