--- src/CMakeLists.txt
@@ -1,14 +1,0 @@
-add_subdirectory(lib-python)
-add_subdirectory(gnssro)
-add_subdirectory(gsi-ncdiag)
-add_subdirectory(marine)
-add_subdirectory(chem)
-if (HAVE_ODB_API)
-add_subdirectory(ODB2)
-endif (HAVE_ODB_API)
-if (odc_FOUND)
-  add_subdirectory(odc)
-endif (odc_FOUND)
-add_subdirectory(ncep)
-add_subdirectory(nsmc)
-add_subdirectory(wrfda-ncdiag)
