+++ src/chem/tropomi_no2_nc2ioda.py
@@ -13,8 +13,13 @@ import netCDF4 as nc
 import numpy as np
 from datetime import datetime, timedelta
 import os
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 import ioda_conv_ncio as iconv
 from collections import defaultdict, OrderedDict
 from orddicts import DefaultOrderedDict
