+++ src/gsi-ncdiag/combine_conv.py
@@ -12,13 +12,10 @@ from collections import defaultdict, OrderedDict
 import datetime as dt
 from pathlib import Path
 
-# Locate src/lib-python/ioda_conv_ncio
-# TODO Replace with setup.py
-IODA_CONV_PATH = "@SCRIPT_LIB_PATH@"
-if Path(IODA_CONV_PATH).is_dir():
-    sys.path.append(IODA_CONV_PATH)
-else:
-    sys.path.append(str((Path(__file__).parent/'..'/'lib-python').resolve()))
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
 
 import ioda_conv_ncio as iconv
 from orddicts import DefaultOrderedDict
