+++ src/gsi-ncdiag/proc_gsi_ncdiag.py
@@ -7,13 +7,10 @@ import time
 from multiprocessing import Pool
 from pathlib import Path
 
-# Locate src/lib-python/gsi
-# TODO Replace with setup.py
-IODA_CONV_PATH = "@SCRIPT_LIB_PATH@"
-if Path(IODA_CONV_PATH).is_dir():
-    sys.path.append(IODA_CONV_PATH)
-else:
-    sys.path.append(str((Path(__file__).parent/'..'/'lib-python').resolve()))
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
 
 import gsi_ncdiag as gsid
