+++ src/marine/cryosat_ice2ioda_DBL.py
@@ -10,28 +10,27 @@
 from __future__ import print_function
 import sys
 import argparse
-import netCDF4 as nc
 from datetime import datetime, timedelta
+import netCDF4 as nc
 import dateutil.parser
 import numpy as np
-import subprocess
 import os
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 from orddicts import DefaultOrderedDict
 import ioda_conv_ncio as iconv
+from read_cryosat_L2 import read_cryosat_L2
 
 
 class Observation(object):
 
     def __init__(self, filename, thin, date, writer):
-        print(date)
-
-        if os.path.exists("cryosat_nc4classic.nc"):
-            os.remove("cryosat_nc4classic.nc")
-
-        subprocess.call('nccopy -k nc7 '+filename[0]+' cryosat_nc4classic.nc', shell=True)
-        self.filename = "cryosat_nc4classic.nc"
+        self.filename = filename[0]
         self.thin = thin
         self.date = date
         self.data = DefaultOrderedDict(lambda: DefaultOrderedDict(dict))
@@ -39,18 +38,35 @@ class Observation(object):
         self._read()
 
     def _read(self):
+        ncd = read_cryosat_L2(self.filename)
+        Data_1Hz = ncd['Data_1Hz']
+        Day = Data_1Hz['Day']
+        Second = Data_1Hz['Second']
+        Micsec = Data_1Hz['Micsec']
+        Time_1Hz = Day*86400+Second+Micsec*1e-6
 
-        ncd = nc.MFDataset(self.filename, aggdim='time_20_ku')
-        time = ncd.variables['time_20_ku'][:]
-        lons = ncd.variables['lon_poca_20_ku'][:]
-        lats = ncd.variables['lat_poca_20_ku'][:]
-        vals = ncd.variables['freeboard_20_ku'][:]
-        qc = ncd.variables['flag_prod_status_20_ku'][:]
-        # get base date for file
-        s = ' '.join(
-            ncd.variables['time_20_ku'].units.split(' ')[2:3])
-        reftime = dateutil.parser.parse(s)
-        ncd.close()
+        Data_20Hz = ncd['Data_20Hz']
+        D_Time = Data_20Hz['D_time_mics']*1e-6
+        Lon0 = Data_20Hz['Lon']*1e-7
+        Lat0 = Data_20Hz['Lat']*1e-7
+        Freeboard0 = Data_20Hz['Freeboard']*1e-3
+        Qflag0 = Data_20Hz['Quality_flag']
+        T1 = np.repeat(Time_1Hz, 20)
+        T2 = T1.reshape(len(Time_1Hz), 20)
+        Time0 = T2+D_Time.data
+        time = Time0.reshape(len(T1))
+        Lon = Lon0.data
+        Lat = Lat0.data
+        FB = Freeboard0.data
+        Qf = Qflag0.data
+        lons = Lon.reshape(len(T1))
+        lats = Lat.reshape(len(T1))
+        vals = FB.reshape(len(T1))
+        qc1 = Qf.reshape(len(T1))
+        qc = qc1.astype(int)
+        # base date for file
+        reftime0 = '2000-01-01 00:00:00'
+        reftime = dateutil.parser.parse(reftime0)
 
         valKey = vName, self.writer.OvalName()
         errKey = vName, self.writer.OerrName()
@@ -73,8 +89,6 @@ class Observation(object):
             self.data[0][locKey][errKey] = 0.1
             self.data[0][locKey][qcKey] = qc[i]
 
-        os.remove("cryosat_nc4classic.nc")
-
 
 vName = "sea_ice_freeboard"
 
@@ -92,7 +106,7 @@ AttrData = {
 def main():
 
     parser = argparse.ArgumentParser(
-        description=('Reads ESA Cryosat-2 Ice and converts into IODA NetCDF')
+        description=('Reads Cryosat-2 Ice from ESA and converts into NetCDF')
     )
 
     required = parser.add_argument_group(title='required arguments')
@@ -122,7 +136,6 @@ def main():
 
     # Read in
     ice = Observation(args.input, args.thin, fdate, writer)
-
     # write them out
     AttrData['date_time_string'] = fdate.strftime("%Y-%m-%dT%H:%M:%SZ")
