+++ src/marine/gmao_obs2ioda.py
@@ -15,8 +15,13 @@ from argparse import ArgumentParser, ArgumentDefaultsHelpFormatter
 import netCDF4 as nc4
 import numpy as np
 from datetime import datetime
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 import ioda_conv_ncio as iconv
 from orddicts import DefaultOrderedDict
