+++ src/marine/hgodas_adt2ioda.py
@@ -12,51 +12,18 @@ import sys
 import argparse
 import netCDF4 as nc
 from datetime import datetime, timedelta
-import dateutil.parser
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 import ioda_conv_ncio as iconv
 from orddicts import DefaultOrderedDict
 
 
-class Observation(object):
-
-    def __init__(self, filename, date, writer):
-
-        self.filename = filename
-        self.date = date
-        self.data = DefaultOrderedDict(lambda: DefaultOrderedDict(dict))
-        self.writer = writer
-        self._read()
-
-    def _read(self):
-
-        ncd = nc.MFDataset(self.filename)
-        time = ncd.variables['time_mjd'][:]
-        lons = ncd.variables['lon'][:]
-        lats = ncd.variables['lat'][:]
-        vals = ncd.variables['adt_xgm2016'][:]
-        units = ncd.variables['time_mjd'].units[-23:-4]
-        reftime = dateutil.parser.parse(units)
-        ncd.close()
-
-        valKey = vName, self.writer.OvalName()
-        errKey = vName, self.writer.OerrName()
-        qcKey = vName, self.writer.OqcName()
-
-        count = 0
-        for i in range(len(lons)):
-
-            count += 1
-            obs_date = reftime + timedelta(days=time[i])
-
-            locKey = lats[i], lons[i], obs_date.strftime("%Y-%m-%dT%H:%M:%SZ")
-            self.data[0][locKey][valKey] = vals[i]
-            self.data[0][locKey][errKey] = 0.1
-            self.data[0][locKey][qcKey] = 0
-
-
-vName = "obs_absolute_dynamic_topography"
+vName = "obs_absolute_dynamic_topography",
 
 locationKeyList = [
     ("latitude", "float"),
@@ -69,21 +36,57 @@ AttrData = {
 }
 
 
-def main():
+class Observation(object):
 
-    # Get command line arguments
+    def __init__(self, filename, date, writer):
+        self.filename = filename
+        self.date = date
+        self.data = DefaultOrderedDict(lambda: DefaultOrderedDict(dict))
+        self.writer = writer
+        self._read()
+
+    def _read(self):
+        ncd = nc.Dataset(self.filename)
+        time = ncd.variables['time'][:]
+        lons = ncd.variables['lon'][:]
+        lats = ncd.variables['lat'][:]
+        hrs = ncd.variables['hr'][:]
+        vals = ncd.variables['val'][:]
+        errs = ncd.variables['err'][:]
+        qcs = ncd.variables['qc'][:]
+        ncd.close()
+
+        base_date = datetime(1970, 1, 1) + timedelta(seconds=int(time[0]))
+
+        valKey = vName, self.writer.OvalName()
+        errKey = vName, self.writer.OerrName()
+        qcKey = vName, self.writer.OqcName()
+
+        for i in range(len(hrs)):
+            # there shouldn't be any bad obs, but just in case remove them all
+            if qcs[i] != 0:
+                continue
+
+            dt = base_date + timedelta(hours=float(hrs[i]))
+            locKey = lats[i], lons[i], dt.strftime("%Y-%m-%dT%H:%M:%SZ")
+            self.data[0][locKey][valKey] = vals[i]
+            self.data[0][locKey][errKey] = errs[i]
+            self.data[0][locKey][qcKey] = qcs[i]
+
+
+def main():
     parser = argparse.ArgumentParser(
         description=(
-            'Reads absolute dynamic topography (ADT) observations'
-            ' from NESDIS file(s) and converts into IODA formatted'
-            ' output files')
+            'Read absolute dynamic topography (ADT) observations'
+            ' file(s) that have already been QCd and thinned for use in'
+            ' Hybrid-GODAS system.')
     )
 
     required = parser.add_argument_group(title='required arguments')
     required.add_argument(
         '-i', '--input',
-        help="RADS observation input file(s)",
-        type=str, nargs='+', required=True)
+        help="name of HGODAS observation input file(s)",
+        type=str, required=True)
     required.add_argument(
         '-o', '--output',
         help="path of ioda output file",
@@ -92,18 +95,18 @@ def main():
         '-d', '--date',
         help="base date for the center of the window",
         metavar="YYYYMMDDHH", type=str, required=True)
-
     args = parser.parse_args()
     fdate = datetime.strptime(args.date, '%Y%m%d%H')
+
     writer = iconv.NcWriter(args.output, locationKeyList)
 
-    # Read in the altimeter
-    altim = Observation(args.input, fdate, writer)
+    # Read in the profiles
+    prof = Observation(args.input, fdate, writer)
 
     # write them out
     AttrData['date_time_string'] = fdate.strftime("%Y-%m-%dT%H:%M:%SZ")
 
-    (ObsVars, LocMdata, VarMdata) = writer.ExtractObsData(altim.data)
+    (ObsVars, LocMdata, VarMdata) = writer.ExtractObsData(prof.data)
     writer.BuildNetcdf(ObsVars, LocMdata, VarMdata, AttrData)
