--- src/marine/hgodas_sst2ioda.py.in
@@ -1,106 +1,0 @@
-#!/usr/bin/env python
-
-#
-# (C) Copyright 2019 UCAR
-#
-# This software is licensed under the terms of the Apache Licence Version 2.0
-# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
-#
-
-from __future__ import print_function
-import sys
-import argparse
-import netCDF4 as nc
-from datetime import datetime, timedelta
-
-sys.path.append("@SCRIPT_LIB_PATH@")
-import ioda_conv_ncio as iconv
-from orddicts import DefaultOrderedDict
-
-
-vName = {
-    'T': "sea_surface_temperature",
-}
-
-locationKeyList = [
-    ("latitude", "float"),
-    ("longitude", "float"),
-    ("datetime", "string")
-]
-
-AttrData = {
-    'odb_version': 1,
-}
-
-
-class Profile(object):
-
-    def __init__(self, filename, date, writer):
-        self.filename = filename
-        self.date = date
-        self.data = DefaultOrderedDict(lambda: DefaultOrderedDict(dict))
-        self.writer = writer
-        self._read()
-
-    def _read(self):
-        ncd = nc.Dataset(self.filename)
-        time = ncd.variables['time'][:]
-        lons = ncd.variables['lon'][:]
-        lats = ncd.variables['lat'][:]
-        hrs = ncd.variables['hr'][:]
-        vals = ncd.variables['val'][:]
-        errs = ncd.variables['err'][:]
-        qcs = ncd.variables['qc'][:]
-        ncd.close()
-
-        base_date = datetime(1970, 1, 1) + timedelta(seconds=int(time[0]))
-
-        valKey = vName['T'], self.writer.OvalName()
-        errKey = vName['T'], self.writer.OerrName()
-        qcKey = vName['T'], self.writer.OqcName()
-
-        count = 0
-        for i in range(len(hrs)):
-            # there shouldn't be any bad obs, but just in case remove them all
-            if qcs[i] != 0:
-                continue
-
-            count += 1
-            dt = base_date + timedelta(hours=float(hrs[i]))
-            locKey = lats[i], lons[i], dt.strftime("%Y-%m-%dT%H:%M:%SZ")
-            self.data[0][locKey][valKey] = vals[i]
-            self.data[0][locKey][errKey] = errs[i]
-            self.data[0][locKey][qcKey] = qcs[i]
-
-
-def main():
-
-    parser = argparse.ArgumentParser(
-        description=('Read CPC Hybrid-GODAS sst files and convert'
-                     ' to IODA format')
-    )
-    parser.add_argument('-i', '--input',
-                        help="name of HGODAS profile input file",
-                        type=str, required=True)
-    parser.add_argument('-o', '--output',
-                        help="name of ioda output file",
-                        type=str, required=True)
-    parser.add_argument('-d', '--date',
-                        help="base date", type=str, required=True)
-    args = parser.parse_args()
-    fdate = datetime.strptime(args.date, '%Y%m%d%H')
-
-    writer = iconv.NcWriter(args.output, locationKeyList)
-
-    # Read in the profiles
-    prof = Profile(args.input, fdate, writer)
-
-    # write them out
-    AttrData['date_time_string'] = fdate.strftime("%Y-%m-%dT%H:%M:%SZ")
-
-    (ObsVars, LocMdata, VarMdata) = writer.ExtractObsData(prof.data)
-    writer.BuildNetcdf(ObsVars, LocMdata, VarMdata, AttrData)
-
-
-if __name__ == '__main__':
-    main()
