+++ src/marine/viirs_modis_oc2ioda.py
@@ -16,8 +16,13 @@ import dateutil.parser
 import numpy as np
 from multiprocessing import Pool
 import os
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 import ioda_conv_ncio as iconv
 
 output_var_names = [
