+++ src/wrfda-ncdiag/proc_wrfda_ncdiag.py
@@ -4,8 +4,13 @@ import os
 import sys
 from multiprocessing import Pool
 import glob
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 import wrfda_ncdiag as wrfdad
