+++ src/wrfda-ncdiag/test_wrfdadiag.py
@@ -2,8 +2,13 @@
 # script to run to test if the WRFDA ncdiag converters are still working
 import sys
 import argparse
+from pathlib import Path
 
-sys.path.append("@SCRIPT_LIB_PATH@")
+IODA_CONV_PATH = Path(__file__).parent/"@SCRIPT_LIB_PATH@"
+if not IODA_CONV_PATH.is_dir():
+    IODA_CONV_PATH = Path(__file__).parent/'..'/'lib-python'
+sys.path.append(str(IODA_CONV_PATH.resolve()))
+
 import wrfda_ncdiag as wrfdad
 
 parser = argparse.ArgumentParser(
