+++ tools/CMakeLists.txt
@@ -1,13 +1,11 @@
 list( APPEND programs
     ${PROJECT_NAME}_lint.sh
+    ${PROJECT_NAME}_comp.sh
 )
 
-SET_TARGETS_DEPS( "${programs}"
+set_targets_deps( "${programs}"
                    ${CMAKE_CURRENT_SOURCE_DIR}
                    ${CMAKE_BINARY_DIR}/bin
                    tools_scripts_deps)
-add_custom_target( tools_scripts ALL DEPENDS ${tools_scripts_deps} )
 
-install (PROGRAMS
-  ${programs}
-  DESTINATION bin)
+install( PROGRAMS ${programs} DESTINATION ${CMAKE_BINARY_DIR} )
